Optimize backlink cache refresh
Receiving backlinks query results from Neuron as single line JSON
strings seems to be rather inefficient (reason still not fully known).
Instead parse the output with jq and output that as a single JSON array
for later usage.

Reverting this to the original method to the upstream seemed to very
degrading to performance.